Organize code structure
