+++ src/components/auth/RouteGuard.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
-import { useCurrentUser } from "@/lib/useCurrentUser";
+import { useCurrentUser } from "@/lib/firebase/useCurrentUser";
 
 interface ProtectedRouteProps {
   requiredRole?: "emisor" | "inversionista";
