+++ src/components/ui/Sidebar.tsx
@@ -20,7 +20,7 @@ import {
   X,
   Calculator,
 } from "lucide-react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/common/utils";
 import { useState } from "react";
 
 const navigationItems = [
