+++ src/components/ui/alert.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/common/utils";
 
 const Alert = React.forwardRef<
   HTMLDivElement,
